+++ tests/testthat/tests.brmsfit-methods.R
@@ -53,7 +53,8 @@ test_that("all S3 methods have reasonable ouputs", {
   # bridge_sampler
   # only test error messages for now
   expect_error(bridge_sampler(fit1), 
-               "Models including prior samples are not usable")
+    "the compiled object from C\\+\\+ code for this model is invalid"
+  )
   
   # coef
   coef1 <- SM(coef(fit1))
@@ -318,11 +319,11 @@ test_that("all S3 methods have reasonable ouputs", {
   expect_equal(nsamples(fit1, incl_warmup = TRUE), 400)
   
   # parnames 
-  expect_equal(parnames(fit1)[c(1, 8, 9, 13, 15, 17, 27, 35, 42, 50, 51)],
+  expect_equal(parnames(fit1)[c(1, 8, 9, 13, 15, 17, 27, 35, 45, 46, 47)],
                c("b_Intercept", "bsp_moExp", "ar[1]", "cor_visit__Intercept__Trt1", 
                  "nu", "simo_moExp1[2]", "r_visit[4,Trt1]", "s_sAge_1[8]", 
                  "prior_sd_visit", "prior_cor_visit", "lp__"))
-  expect_equal(parnames(fit2)[c(1, 4, 6, 7, 9, 71, 129)],
+  expect_equal(parnames(fit2)[c(1, 4, 6, 7, 9, 71, 127)],
                c("b_a_Intercept", "b_b_Age", "sd_patient__b_Intercept",
                  "cor_patient__a_Intercept__b_Intercept", 
                  "r_patient__a[1,Intercept]", "r_patient__b[4,Intercept]",
@@ -484,9 +485,8 @@ test_that("all S3 methods have reasonable ouputs", {
   # prior_samples
   prs1 <- prior_samples(fit1)
   prior_names <- c(
-    "temp_Intercept", "temp_sigma_Intercept", "b_Intercept", 
-    "b_sigma_Intercept", "sds_sAge_1", "nu", "sd_visit", "b", "bsp", 
-    paste0("simo_moExp1[", 1:4, "]"), "b_sigma", "cor_visit"
+    "b", "bsp", paste0("simo_moExp1[", 1:4, "]"), 
+    "sds_sAge_1", "b_sigma", "nu", "sd_visit", "cor_visit"
   )
   expect_equal(colnames(prs1), prior_names)